backend/requirements.txt
django==3.2.3
django-filter==21.1
python-dotenv==0.21.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
djangorestframework==3.14.0
djoser==2.1.0
Pillow==9.2.0
reportlab==3.6.11